chore: relocate component files
